Move EmbeddedFrame to dedicated directory
