+++ components/SegmentCreationWizard.js
@@ -1,5 +1,5 @@
 import Card from "@components/Card/Card"
-import EmbeddedFrame from "@components/EmbeddedFrame"
+import EmbeddedFrame from "@components/EmbeddedFrame/EmbeddedFrame"
 
 export default function SegmentCreationWizard({
   setSegments,
